requirements.txt
greenlet==3.2.4
numpy==2.3.4
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
imbalanced-learn==0.12.3